Extract diff baseline selector helpers
